# fix: ensure env command is package manager specific