@agdimech 2023-05-05
Not sure why jsii cannot be found as locally the `node_modules/.bin` folder is added to the PATH.